updated stepfunctions example to reflect format from cdk init template
