+++ python/stepfunctions/requirements.txt
@@ -1,5 +1,1 @@
-aws-cdk.cdk
-
-aws-cdk.aws-stepfunctions
-aws-cdk.aws-stepfunctions-tasks
-
+-e .
